CodebaseExplainer/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=1.10.0
requests>=2.31.0
PyGithub>=2.1.3
anthropic>=0.51.0
gitpython>=3.1.0
nbformat>=5.0.0
nbconvert>=6.0.0
pytest>=6.2.0
